requirements.txt
Flask==1.1.2
gunicorn==20.1.0
html5lib==1.1
imbalanced-learn==0.8.1
Jinja2==2.11.2
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.2
nltk==3.5
numpy==1.19.2
numpydoc==1.1.0
pandas==1.1.3
pip==21.3.1
py==1.9.0
pycrf==0.0.1
pycurl==7.43.0.6
pytest==6.2.5
python-crfsuite==0.9.7
python-language-server==0.35.1
regex==2020.10.15
requests==2.24.0
requests-file==1.5.1
requests-oauthlib==1.3.0
scikit-image==0.18.0
scikit-learn==1.0.2
scipy==1.7.0
sklearn-crfsuite==0.3.6
spacy==3.0.7
unicodecsv==0.14.1
xgboost==1.4.2
